requirements.txt
opencv-python
opencv-contrib-python
mediapipe==0.10.14
face_recognition
dlib-bin
numpy
python-dotenv
langchain
langchain-community
langchain-text-splitters
sentence-transformers
faiss-cpu
pymupdf
google-genai